# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Kato Backend Core API
  version: 0.3.3
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/bootstrap-tokens:
    post:
      summary: Create one-time bootstrap token
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum:
                    - frontend-edge
                    - subscription-edge
                    - proxy-node
                    - transit-relay
                name:
                  type: string
                ttlSeconds:
                  type: integer
                  default: 900
                resourceId:
                  type:
                    - string
                    - "null"
      responses:
        "201":
          description: Bootstrap token created
  /api/v1/auth/login:
    post:
      summary: Login with an admin username and password through a paired panel frontend
      security:
        - FrontendToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Admin session created
  /api/v1/auth/session:
    get:
      summary: Read current admin session
      security:
        - FrontendToken: []
          AdminSession: []
      responses:
        "200":
          description: Current admin session
  /api/v1/auth/logout:
    post:
      summary: Revoke current admin session
      security:
        - FrontendToken: []
          AdminSession: []
      responses:
        "200":
          description: Session revoked
...
  /api/v1/agents/register:
    post:
      summary: Register an Edge or Agent with a bootstrap token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bootstrapToken
              properties:
                bootstrapToken:
                  type: string
                agentVersion:
                  type: string
                hostname:
                  type: string
                capabilities:
                  type: object
      responses:
        "201":
          description: Agent registered
...
components:
...
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
    FrontendToken:
      type: apiKey
      in: header
      name: X-Frontend-Token
    AdminSession:
      type: http
      scheme: bearer
    AgentBearer:
      type: http
      scheme: bearer